+++ .gitignore
@@ -27,7 +27,7 @@ share/python-wheels/
 MANIFEST
 
 # PyInstaller
-#  Usually these files are written by a python script from a template_app
+#  Usually these files are written by a python script from a template_module
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
 *.manifest
 *.spec
