+++ backend/modules/template_module/adapters/repositories/sqlalchemy/consts.py
