+++ backend/modules/template_module/domain/exceptions.py
