+++ backend/modules/template_module/domain/ports/exceptions.py
