+++ backend/modules/template_module/domain/ports/repository.py
