+++ backend/modules/template_module/domain/value_objects.py
