+++ backend/tests/integrated/template_module/conftest.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 import pytest
 
-from apps.template_app.domain.entities import Template as TemplateEntity
+from modules.template_module.domain.entities import Template as TemplateEntity
 from .factories import FakeTemplateRepository, FakeTemplateUnitOfWork
